+++ bin/bin.go
@@ -158,3 +158,37 @@ func Rwd() string {
 	}
 	return appRuningWorkDir
 }
+
+// 获取字符串格式化
+// [[k,  v]]
+func FormatStr(d string, ss ...[][]string) string {
+	if d == "" {
+		// 4 个空格
+		d = "   "
+	}
+	bit := d[0:1]
+
+	// 或者最大长度
+	maxLen := 0
+	for _, sg := range ss {
+		for _, s := range sg {
+			kLen := len(s[0])
+			if kLen > maxLen {
+				maxLen = kLen
+			}
+		}
+	}
+
+	maxLen += len(d)
+
+	// 格式化
+	var contents string
+	for _, sg := range ss {
+		for _, s := range sg {
+			ss1 := s[0] + strings.Repeat(bit, maxLen - len(s[0])) + s[1] + "\n"
+			contents += ss1
+		}
+	}
+
+	return contents
+}